+++ pages/manageFormula.php
@@ -8,7 +8,9 @@ require_once(__ROOT__.'/inc/settings.php');
 require_once(__ROOT__.'/func/labelMap.php');
 require_once(__ROOT__.'/func/get_formula_notes.php');
 
-
+if($_POST['update_rating'] == '1' && $_POST['fid'] && $_POST['score']){
+	mysqli_query($conn,"UPDATE formulasMetaData SET rating = '".$_POST['score']."' WHERE id = '".$_POST['fid']."'");
+}
 
 //EXCLUDE/INCLUDE INGREDIENT
 if($_POST['action'] == 'excIng' && $_POST['ingID']){