+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New:
 ----------------------
+- A rating option has been added so each formula can be rated
 - Added the option to attach documents in formulas
 - Formula making view improvements